Image_Captioning/requirements.txt
tensorflow==2.1.0
opencv-python
matplotlib
sklearn
re
numpy
os
json
time
pillow
glob
nltk